# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 'Autpost API'
  description: 'API documentation for the Autpost application'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:9527/api'
    description: 'Local development server'
paths:
...
  /recurring-todos:
    get:
      tags:
        - RecurringTodos
      summary: 'Get all recurring todos for the authenticated user'
      operationId: abf7b35772940fdce25fdb3c4fd664e2
      parameters:
        -
          name: state
          in: query
          required: false
          schema:
            type: string
            enum:
              - active
              - paused
        -
          name: include_deleted
          in: query
          required: false
          schema:
            type: boolean
        -
          name: page
          in: query
          required: false
          schema:
            type: integer
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'List of recurring todos with pagination'
        401:
          description: Unauthenticated
      security:
        -
          bearerAuth: []
    post:
      tags:
        - RecurringTodos
      summary: 'Create a new recurring todo'
      operationId: e943f2135019d4ca70a3cb325b4319be
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - interval
                - interval_unit
                - start_time
              properties:
                title:
                  type: string
                  maxLength: 255
                note:
                  type: string
                  nullable: true
                interval:
                  type: integer
                  minimum: 1
                interval_unit:
                  type: string
                  enum:
                    - day
                    - week
                    - month
                    - year
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                  nullable: true
                state:
                  type: string
                  enum:
                    - active
                    - paused
                  nullable: true
              type: object
      responses:
        201:
          description: 'Recurring todo created successfully'
        401:
          description: Unauthenticated
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
  '/recurring-todos/{id}':
    get:
      tags:
        - RecurringTodos
      summary: 'Get a single recurring todo'
      operationId: 6042d2e4c12abc01762460ba8ebfe13c
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo details'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - RecurringTodos
      summary: 'Update a recurring todo'
      operationId: 27c0063e8191f3b8388e262f4c30f3b0
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  maxLength: 255
                note:
                  type: string
                  nullable: true
                interval:
                  type: integer
                  minimum: 1
                interval_unit:
                  type: string
                  enum:
                    - day
                    - week
                    - month
                    - year
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                  nullable: true
                state:
                  type: string
                  enum:
                    - active
                    - paused
                  nullable: true
              type: object
      responses:
        200:
          description: 'Recurring todo updated successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - RecurringTodos
      summary: 'Soft delete a recurring todo'
      operationId: bcfce21aae1e9ffb971e8fc222b3a0c2
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo deleted successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
  '/recurring-todos/{id}/restore':
    post:
      tags:
        - RecurringTodos
      summary: 'Restore a soft-deleted recurring todo'
      operationId: 716f8ceff48bc100e7fd50ec7e17fa94
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo restored successfully'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
  '/recurring-todos/{id}/force':
    delete:
      tags:
        - RecurringTodos
      summary: 'Permanently delete a recurring todo'
      operationId: 1ef010273412bef9ab3b5a6c52f4318e
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo permanently deleted'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
  '/recurring-todos/{id}/pause':
    post:
      tags:
        - RecurringTodos
      summary: 'Pause a recurring todo'
      operationId: a7f2399ced54734b3ff51483fd8f774f
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo paused'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
  '/recurring-todos/{id}/resume':
    post:
      tags:
        - RecurringTodos
      summary: 'Resume a recurring todo'
      operationId: ec7765eea74812680b5144cf80c61070
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Recurring todo resumed'
        401:
          description: Unauthenticated
        404:
          description: 'Recurring todo not found'
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Enter your Sanctum token'
      bearerFormat: token
      scheme: bearer
...